Add files via upload
Scenario
A company wants to build a three-tier web app with maximum security 
against internet vulnerabilities such as DDOS attack behind an application load balancer. 
The solution must be highly available, and scalable to accommodate to the volume of requests. 
The company also wants to monitor the performance of their applications to enable
compliance auditing, security analysis, and resource change tracking.